Restructure files + more improvements
